# api.yaml
openapi: 3.1.0
info:
  title: Project & Task Management API
  description: API for managing companies, users, roles, projects, members, tasks, comments, history, attachments, deadlines, reminders, notifications, and time tracking.
  version: 1.0.0
servers:
  - url: "http://localhost:8080"
security:
  - bearerAuth: []
...
paths:
...
  /tasks/{taskId}:
    get:
      summary: Get task by ID
      tags: [Tasks]
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update task
      tags: [Tasks]
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


    delete:
      summary: Delete task
      tags: [Tasks]
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        204:
          description: Deleted
...
  /tasks/{taskId}/reminders:
    get:
      summary: Get reminders for task
      tags: [Task Reminders]
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskReminder"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Add reminder to task
      tags: [Task Reminders]
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskReminderCreateRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskReminder"

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}/reminders/{reminderId}:
    delete:
      summary: Delete reminder
      tags: [Task Reminders]
      parameters:
        - $ref: "#/components/parameters/taskId"
        - in: path
          name: reminderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Deleted
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    taskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ OPEN, IN_PROGRESS, DONE, CANCELED]
        priority:
          type: string
          enum: [ LOW, MEDIUM, HIGH, CRITICAL ]
        created_by:
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        due_date:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assigned_to:
          type: string
          format: uuid
        due_date:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    TaskReminder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        remind_at:
          type: string
          format: date-time
        message:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskReminderCreateRequest:
      type: object
      required: [ remind_at, message, created_by ]
      properties:
        remind_at:
          type: string
          format: date-time
        message:
          type: string
        created_by:
          type: string
          format: uuid
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string   